REF: share MaskedArray methpds
