+++ pandas/core/arrays/numeric.py
@@ -16,7 +16,6 @@ from pandas._typing import (
     Dtype,
     DtypeObj,
 )
-from pandas.compat.numpy import function as nv
 from pandas.errors import AbstractMethodError
 
 from pandas.core.dtypes.common import (
@@ -211,40 +210,3 @@ class NumericArray(BaseMaskedArray):
         return cls._from_sequence(scalars, dtype=dtype, copy=copy)
 
     _HANDLED_TYPES = (np.ndarray, numbers.Number)
-
-    def __neg__(self):
-        return type(self)(-self._data, self._mask.copy())
-
-    def __pos__(self):
-        return self.copy()
-
-    def __abs__(self):
-        return type(self)(abs(self._data), self._mask.copy())
-
-    def round(self: T, decimals: int = 0, *args, **kwargs) -> T:
-        """
-        Round each value in the array a to the given number of decimals.
-
-        Parameters
-        ----------
-        decimals : int, default 0
-            Number of decimal places to round to. If decimals is negative,
-            it specifies the number of positions to the left of the decimal point.
-        *args, **kwargs
-            Additional arguments and keywords have no effect but might be
-            accepted for compatibility with NumPy.
-
-        Returns
-        -------
-        NumericArray
-            Rounded values of the NumericArray.
-
-        See Also
-        --------
-        numpy.around : Round values of an np.array.
-        DataFrame.round : Round values of a DataFrame.
-        Series.round : Round values of a Series.
-        """
-        nv.validate_round(args, kwargs)
-        values = np.round(self._data, decimals=decimals, **kwargs)
-        return type(self)(values, self._mask.copy())
